QoSAwareScheduler: fix integer division in PF score computation
Bug: In prepareSchedule(), the PF score formula computed
availableBytes / availableBlocks where both operands are unsigned int.
This truncated the result to an integer before the subsequent float
division by pfRate_, losing precision and producing coarser scores
than intended.

Fix: Cast availableBytes to double before the division.

+++ src/simu5g/stack/mac/scheduling_modules/QoSAwareScheduler.cc
@@ -104,7 +104,7 @@ void QoSAwareScheduler::prepareSchedule()
         if (pfRate_[cid] < scoreEpsilon_)
             s = qosWeight / scoreEpsilon_;
         else if (availableBlocks > 0)
-            s = qosWeight * ((availableBytes / availableBlocks) / pfRate_[cid])
+            s = qosWeight * (static_cast<double>(availableBytes) / availableBlocks / pfRate_[cid])
                 + uniform(getEnvir()->getRNG(0), -scoreEpsilon_ / 2.0, scoreEpsilon_ / 2.0);
         else
             s = 0.0;
